Change color into f32 struct.

+++ src/render/fill.rs
@@ -3,6 +3,6 @@ use crate::framebuffer::PPM;
 
 pub fn fill(buf: &mut (impl PPM + ?Sized), w: usize, h: usize, color: Color) {
     for i in 0..w * h {
-        buf.ppm_set_i(i as usize, color);
+        buf.ppm_set_i(i, color);
     }
 }
